moved remnant basic test to Settings class
where it belongs
