+++ UnitTests.Core/Settings.cs
@@ -14,6 +14,17 @@
     public class SettingsTests
     {
         [TestMethod]
+        public void Can_load_a_fake_xml_configuration()
+        {
+            var logger = Substitute.For<ILogEvents>();
+
+            var settings = TestHelpers.LoadConfigFromResourceFile("NoOp.policies", logger);
+            var level = settings.LogLevel;
+
+            Assert.AreEqual(LogLevel.Diagnostic, level);
+        }
+
+        [TestMethod]
         [ExpectedException(typeof(FileNotFoundException))]
         public void Raise_error_for_non_existing_file()
         {
