+++ UnitTests.Core/TestHelpers.cs
@@ -14,7 +14,7 @@
         {
             try
             {
-                var thisAssembly = Assembly.GetAssembly(typeof(Basics));
+                var thisAssembly = Assembly.GetAssembly(typeof(TestHelpers));
                 var stream = thisAssembly.GetManifestResourceStream("UnitTests.Core.ConfigurationsForTests." + resourceName);
                 var textStream = new StreamReader(stream);
                 return textStream.ReadToEnd();
